Move plan2svg file to separate dir
