+++ ydb/library/plan2svg/ya.make
@@ -1,0 +1,7 @@
+LIBRARY()
+
+SRCS(
+    plan2svg.cpp
+)
+
+END()
